feat: database and table setup (#15)
# Description

Adds code to define three basic tables. 

## Guilds

Every guild will have some basic global properties. Currently, the only
property globally is the ID of the text channel to put the quotes into.

## Statements and Exchanges

Every entry in the quote board will be an "exchange" composed of one or
more statements and an optional context for the overall exchange.